more inputbox checking, vcb depths now float
depth factors for pockets.centerline,fold now can be floating point.
all options menus use rescue on inputbox calls
all plunge cdia cdepth etc are now length objects
various getattribute calls now return numbers instead of strings

+++ Plugins/Phlatboyz/tools/CenterLineTool.rb
@@ -111,7 +111,7 @@ module PhlatScript
      # VCB
     def onUserText(text,view)
       begin
-         value = text.to_i
+         value = text.to_f
       rescue
          # Error parsing the text
          UI.beep
